Use direct imports instead of Config class
- Removed the Config class from the configuration module.
- Updated the main code to directly import and use configuration constants
- Change filename from config.py to constants.py
