+++ main.py
@@ -15,29 +15,27 @@ from chat_logging import (
     generate_csv
 )
 
-from config import Config
-Config.check_required_env_vars()
 
-# Use configuration values
-IS_DEVELOPER_VIEW = Config.IS_DEVELOPER_VIEW
-ON_LOCALHOST = Config.ON_LOCALHOST
-MCM_URL = Config.MCM_URL
-SKILL_NAME_TO_MCM_URL = Config.SKILL_NAME_TO_MCM_URL
-COGNITO_DOMAIN = Config.COGNITO_DOMAIN
-URL_CODE = Config.URL_CODE
-ACCESS_TOKEN = Config.ACCESS_TOKEN
-REDIRECT_URL = Config.REDIRECT_URL
-CLIENT_ID = Config.CLIENT_ID
-CLIENT_SECRET = Config.CLIENT_SECRET
-LOGIN_URL = Config.LOGIN_URL
-GET_ACCESS_TOKEN_URL = Config.GET_ACCESS_TOKEN_URL
-GET_USER_INFO_URL = Config.GET_USER_INFO_URL
-USER_NAME = Config.USER_NAME
-USERNAME = Config.USERNAME
-USER_EMAIL = Config.USER_EMAIL
+from constants import (
+    IS_DEVELOPER_VIEW,
+    ON_LOCALHOST,
+    MCM_URL,
+    SKILL_NAME_TO_MCM_URL,
+    COGNITO_DOMAIN,
+    URL_CODE,
+    ACCESS_TOKEN,
+    REDIRECT_URL,
+    CLIENT_ID,
+    CLIENT_SECRET,
+    LOGIN_URL,
+    GET_ACCESS_TOKEN_URL,
+    GET_USER_INFO_URL,
+    USER_NAME,
+    USERNAME,
+    USER_EMAIL
+)
 
 app = FastAPI()
-
 
 @app.get("/")
 def read_main():
@@ -49,9 +47,9 @@ def get_access_token_and_user_info(url_code):
     URL_CODE = url_code
     access_token_data = {
         "grant_type": "authorization_code",
-        "client_id": Config.CLIENT_ID,
+        "client_id": CLIENT_ID,
         "code": URL_CODE,
-        "redirect_uri": Config.REDIRECT_URL,
+        "redirect_uri": REDIRECT_URL,
     }
     access_token_headers = {
         "Authorization": "Basic "
@@ -70,18 +68,18 @@ def get_access_token_and_user_info(url_code):
     try:
         # Get Access Token
         response = requests.post(
-            Config.GET_ACCESS_TOKEN_URL, data=access_token_data, headers=access_token_headers
+            GET_ACCESS_TOKEN_URL, data=access_token_data, headers=access_token_headers
         )
         access_token = response.json()["access_token"]
-        Config.ACCESS_TOKEN = access_token
+        ACCESS_TOKEN = access_token
 
         # Get User Info
         response = requests.get(
-            Config.GET_USER_INFO_URL, headers=get_user_info_header(access_token)
+            GET_USER_INFO_URL, headers=get_user_info_header(access_token)
         ).json()
-        Config.USER_NAME = response["name"]
-        Config.USERNAME = response["username"]
-        Config.USER_EMAIL = response["email"]
+        USER_NAME = response["name"]
+        USERNAME = response["username"]
+        USER_EMAIL = response["email"]
         return True
     except Exception as e:
         print(str(e))
@@ -92,7 +90,7 @@ def get_access_token_and_user_info(url_code):
 def get_mcm_response(question: str) -> str:
     try:
         response = httpx.post(
-            Config.MCM_URL,
+            MCM_URL,
             json={"question": question, "api_key": mcm_api_key.value, "Episodic_Knowledge": {}},
             timeout=timeout_secs.value,
         )
@@ -164,7 +162,7 @@ with gr.Blocks() as ivy_main_page:
         if not get_access_token_and_user_info(url_code):
             # (TODO): Redirect to Login page or display Error page
             return "An Error Occurred"
-        return f"# Welcome to Ivy Chatbot, {Config.USERNAME}"
+        return f"# Welcome to Ivy Chatbot, {USER_NAME}"
 
     
     def update_user_message(user_message, history):
@@ -179,7 +177,7 @@ with gr.Blocks() as ivy_main_page:
             yield history
         # Log to DynamoDB every interaction here
         log_chat_history(
-            Config.USERNAME, Config.ACCESS_TOKEN, history[-1][0], history[-1][1], "no_reaction"
+            USERNAME, ACCESS_TOKEN, history[-1][0], history[-1][1], "no_reaction"
         )
 
     def log_commended_response(history):
@@ -187,7 +185,7 @@ with gr.Blocks() as ivy_main_page:
             return
         response = history[-1][1]
         question = history[-1][0]
-        log_chat_history(Config.USERNAME, Config.ACCESS_TOKEN, question, response, "liked")
+        log_chat_history(USERNAME, ACCESS_TOKEN, question, response, "liked")
         gr.Info("Saved successfully!")
 
     def log_disliked_response(history):
@@ -195,7 +193,7 @@ with gr.Blocks() as ivy_main_page:
             return
         response = history[-1][1]
         question = history[-1][0]
-        log_chat_history(Config.USERNAME, Config.ACCESS_TOKEN, question, response, "disliked")
+        log_chat_history(USERNAME, ACCESS_TOKEN, question, response, "disliked")
         gr.Info("Saved successfully!")
 
     def log_flagged_response(history):
@@ -203,7 +201,7 @@ with gr.Blocks() as ivy_main_page:
             return
         response = history[-1][1]
         question = history[-1][0]
-        log_chat_history(Config.USERNAME, Config.ACCESS_TOKEN, question, response, "flagged")
+        log_chat_history(USERNAME, ACCESS_TOKEN, question, response, "flagged")
         gr.Info("Saved successfully!")
 
     def chat_liked_or_disliked(history, data: gr.LikeData):
@@ -215,11 +213,11 @@ with gr.Blocks() as ivy_main_page:
             log_disliked_response([[question, response]])
 
     def handle_download_click():
-        filepath = generate_csv(Config.USERNAME, Config.ACCESS_TOKEN)
+        filepath = generate_csv(USERNAME, ACCESS_TOKEN)
         return filepath if filepath else None
 
     def update_skill(skill_name):
-        Config.MCM_URL = Config.SKILL_NAME_TO_MCM_URL[skill_name]
+        MCM_URL = SKILL_NAME_TO_MCM_URL[skill_name]
         return []
 
     ivy_main_page.load(on_page_load, None, [welcome_msg])
